# Conflicts:
#	src/test/java/ru/itmo/idu/geometry/GeometryUtilsTest.java

+++ src/main/java/ru/itmo/idu/geometry/ProjectionUtils.java
@@ -173,7 +173,7 @@ public class ProjectionUtils {
      * Builds a correct round buffer around given point. Suitable for making circular areas on the map look really circular
      * (as a circle in WGS84 coordinates will not be a circle on the map but an ellipse)
      */
-    public static Geometry makePointBuffer(Coordinate coord, int meters) {
+    public static Geometry makePointBuffer(Coordinate coord, double meters) {
         Point point = makePoint(coord);
         return bufferProjected(point, meters);
     }
